// src/backend/laya-calibration.generated.ts
export const LAYA_CALIBRATIONS: Partial<Record<LayaModel, LayaCalibration>> = {
  "english": {
    "version": "cal-20260923",
    "featureVersion": "lf-1",
    "tierWeights": [
      [
        0.342294,
        -0.789425,
        -1.057255,
        -3.643384,
        0.745261,
        0.169209,
        0.116592,
        -0.192162,
        -0.449353,
        -0.170045,
        -0.33244,
        0.161432,
        -0.101868,
        1.931226
      ],
      [
        -0.200337,
        -0.09063,
        -0.221313,
        -1.398529,
        0.387938,
        0.089587,
        -0.338667,
        0.041642,
        0.129924,
        -0.025369,
        -0.427045,
        0.105305,
        -0.014405,
        1.375769
      ],
      [
        -0.142301,
        0.879711,
        1.278966,
        5.041918,
        -1.133279,
        -0.258859,
        0.221984,
        0.150516,
        0.319408,
        0.195393,
        0.759478,
        -0.266724,
        0.116376,
        -2.044495
      ]
    ],
    "demandWeights": [
      0.950895,
      0.587091,
      2.050732,
      0.573524,
      -0.373571,
      -0.114759,
      -0.01016,
      0.06985,
      0.201619,
      0.089392,
      0.253768,
      -0.159367,
      -0.136447,
      4.520457
    ],
    "fit": {
      "samples": 440,
      "groups": 327,
      "lambda": 0.01,
      "eps": 0.1,
      "cv_agree": 0.57,
      "cv_under": 0.039,
      "cv_over": 0.391,
      "cv_plan_agree": 0.573,
      "cv_plan_under": 0.041,
      "sources": "{\"corpus:main\":240,\"corpus:subagent\":80,\"corpus2:subagent\":30,\"corpus2:main\":90}"
    }
  },
  "typed-decisions": {
    "version": "cal-20260923",
    "featureVersion": "lf-1",
    "tierWeights": [
      [
        0.589215,
        -0.621892,
        -1.080688,
        -3.14416,
        0.477386,
        0.21039,
        0.047444,
        -0.244651,
        -0.718719,
        -0.561165,
        -0.547908,
        0.308216,
        -0.088207,
        2.010245
      ],
      [
        -0.060974,
        0.066704,
        0.534255,
        0.406894,
        0.334491,
        0.323906,
        -0.403995,
        0.089698,
        0.128498,
        0.11687,
        -0.86378,
        0.117878,
        -0.008458,
        0.998198
      ],
      [
        -0.528044,
        0.555378,
        0.546224,
        2.736897,
        -0.811952,
        -0.534351,
        0.356476,
        0.154927,
        0.590172,
        0.444227,
        1.411622,
        -0.426055,
        0.096617,
        -2.070892
      ]
    ],
    "demandWeights": [
      0.565831,
      0.688888,
      1.25011,
      0.80803,
      -0.392026,
      -0.254781,
      0.054087,
      0.151156,
      0.540485,
      0.39081,
      0.438862,
      -0.259738,
      -0.164322,
      3.445108
    ],
    "fit": {
      "samples": 466,
      "groups": 353,
      "lambda": 0.03,
      "eps": 0.1,
      "cv_agree": 0.536,
      "cv_under": 0.032,
      "cv_over": 0.431,
      "cv_plan_agree": 0.539,
      "cv_plan_under": 0.034,
      "sources": "{\"corpus:main\":240,\"corpus:subagent\":80,\"corpus2:subagent\":30,\"corpus2:main\":90,\"session:main\":24,\"session:subagent\":2}"
    }
  },
  "multilingual": {
    "version": "cal-20260923",
    "featureVersion": "lf-1",
    "tierWeights": [
      [
        -0.118942,
        0.236195,
        0.121061,
        -3.51215,
        0.18317,
        0.131482,
        0.065191,
        -0.194018,
        -0.20297,
        -0.085736,
        0.038718,
        0.506254,
        0.414392,
        2.794194
      ],
      [
        0.017721,
        -0.058392,
        -0.207678,
        -0.737857,
        -0.204963,
        0.142698,
        0.021637,
        0.0264,
        -0.103407,
        0.126289,
        -0.039753,
        0.210166,
        0.326163,
        0.895198
      ],
      [
        0.101221,
        -0.177803,
        0.086616,
        4.250012,
        0.021793,
        -0.27418,
        -0.086829,
        0.167617,
        0.306376,
        -0.040554,
        0.001035,
        -0.716421,
        -0.740553,
        -2.748
      ]
    ],
    "demandWeights": [
      0.07739,
      -0.158793,
      -0.3076,
      2.40288,
      -0.054386,
      -0.086246,
      -0.020926,
      0.090035,
      0.125275,
      0.006833,
      -0.028381,
      -0.367625,
      -0.402498,
      -0.021372
    ],
    "fit": {
      "samples": 440,
      "groups": 327,
      "lambda": 0.01,
      "eps": 0.1,
      "cv_agree": 0.477,
      "cv_under": 0.052,
      "cv_over": 0.47,
      "cv_plan_agree": 0.482,
      "cv_plan_under": 0.052,
      "sources": "{\"corpus:main\":240,\"corpus:subagent\":80,\"corpus2:subagent\":30,\"corpus2:main\":90}"
    }
  }
};